Remove UID based dependencies
Over the last base data updates, the uid's have changed. These UIDs were used for hard coded uid based replacements. This leads to errors, because not the expected datasets were addressed and changed. So the replacements were done in the base data, the replacements are no longer required and removed.